+++ sdk/dependencymap/arm-dependencymap/samples-dev/discoverySourcesDeleteSample.ts
@@ -1,16 +1,15 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import { DependencyMapClient } from "@azure/arm-dependencymap";
+import { DefaultAzureCredential } from "@azure/identity";
+
 /**
  * This sample demonstrates how to delete a DiscoverySourceResource
  *
  * @summary delete a DiscoverySourceResource
- * x-ms-original-file: 2025-01-31-preview/DiscoverySources_Delete.json
+ * x-ms-original-file: 2025-07-01-preview/DiscoverySources_Delete.json
  */
-
-import { DependencyMapClient } from "@azure/arm-dependencymap";
-import { DefaultAzureCredential } from "@azure/identity";
-
 async function discoverySourcesDeleteGeneratedByMaximumSetRule(): Promise<void> {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "D6E58BDB-45F1-41EC-A884-1FC945058848";