+++ sdk/dependencymap/arm-dependencymap/samples/v1-beta/javascript/mapsGetDependencyViewForFocusedMachineSample.js
@@ -8,7 +8,7 @@ const { DefaultAzureCredential } = require("@azure/identity");
  * This sample demonstrates how to get dependency map of single machine
  *
  * @summary get dependency map of single machine
- * x-ms-original-file: 2025-01-31-preview/Maps_GetDependencyViewForFocusedMachine.json
+ * x-ms-original-file: 2025-07-01-preview/Maps_GetDependencyViewForFocusedMachine.json
  */
 async function mapsGetDependencyViewForFocusedMachineGeneratedByMaximumSetRule() {
   const credential = new DefaultAzureCredential();