+++ sdk/dependencymap/arm-dependencymap/samples/v1-beta/javascript/mapsGetSample.js
@@ -8,7 +8,7 @@ const { DefaultAzureCredential } = require("@azure/identity");
  * This sample demonstrates how to get a MapsResource
  *
  * @summary get a MapsResource
- * x-ms-original-file: 2025-01-31-preview/Maps_Get.json
+ * x-ms-original-file: 2025-07-01-preview/Maps_Get.json
  */
 async function mapsGetGeneratedByMaximumSetRule() {
   const credential = new DefaultAzureCredential();