+++ sdk/dependencymap/arm-dependencymap/samples/v1-beta/typescript/src/mapsGetConnectionsWithConnectedMachineForFocusedMachineSample.ts
@@ -8,7 +8,7 @@ import { DefaultAzureCredential } from "@azure/identity";
  * This sample demonstrates how to get network connections between machines
  *
  * @summary get network connections between machines
- * x-ms-original-file: 2025-01-31-preview/Maps_GetConnectionsWithConnectedMachineForFocusedMachine.json
+ * x-ms-original-file: 2025-07-01-preview/Maps_GetConnectionsWithConnectedMachineForFocusedMachine.json
  */
 async function mapsGetConnectionsWithConnectedMachineForFocusedMachineGeneratedByMaximumSetRule(): Promise<void> {
   const credential = new DefaultAzureCredential();