+++ sdk/dependencymap/arm-dependencymap/samples/v1-beta/typescript/src/mapsListBySubscriptionSample.ts
@@ -8,7 +8,7 @@ import { DefaultAzureCredential } from "@azure/identity";
  * This sample demonstrates how to list MapsResource resources by subscription ID
  *
  * @summary list MapsResource resources by subscription ID
- * x-ms-original-file: 2025-01-31-preview/Maps_ListBySubscription.json
+ * x-ms-original-file: 2025-07-01-preview/Maps_ListBySubscription.json
  */
 async function mapsListBySubscriptionGeneratedByMaximumSetRule(): Promise<void> {
   const credential = new DefaultAzureCredential();